Add batch dependencies to all deployments if the batch extension is enabled.
